% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_package.R, R/dt_to_parquet.R
\name{add_package}
\alias{add_package}
\title{adds library(<package>) call to file here::here("R/packages.R")}
\usage{
add_package(package, open_file = F)

add_package(package, open_file = F)
}
\arguments{
\item{package}{Unquoeted symbol}

\item{open_file}{Should "R/packages.R" be opened afterwards?}
}
\description{
adds library(<package>) call to file here::here("R/packages.R")

saves a data.table for a parquet filesystem
}
\keyword{drake}
\keyword{parquet}
