% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_from_ags.R
\name{bl_from_ags}
\alias{bl_from_ags}
\title{Get bl from amtlicher Gemeindeschluessel}
\usage{
bl_from_ags(ags)
}
\arguments{
\item{ags}{Amtlicher Gemeindeschluessel, with leading zero. Only first 2 chars relevant}
}
\description{
Get bl from amtlicher Gemeindeschluessel
}
\examples{
bl_from_ags('03254021')
}
\keyword{ags}
