% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pattern_in_text_files.R
\name{find_pattern_in_text_files}
\alias{find_pattern_in_text_files}
\title{Find a pattern within a directory of R files}
\usage{
find_pattern_in_text_files(pattern, folder = here::here())
}
\arguments{
\item{pattern}{The pattern}

\item{folder}{Folder where to search}
}
\description{
Yes, find a pattern
}
\examples{
find_pattern_in_text_files("mean")
}
\keyword{cats}
