% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadd_fun.R
\name{loadd_fun}
\alias{loadd_fun}
\title{drake::loadds all arguments of funtion
Common usecase: on a make_X function}
\usage{
loadd_fun(fun)
}
\arguments{
\item{fun}{function, whose arguments we want to loadd}
}
\description{
drake::loadds all arguments of funtion
Common usecase: on a make_X function
}
\keyword{drake}
