% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chunks.R
\name{make_chunks}
\alias{make_chunks}
\title{Makes approx equal sized chunks of a vector}
\usage{
make_chunks(x, chunksize)
}
\arguments{
\item{x}{Vector to chunk}

\item{chunksize}{Size of chunk}
}
\description{
Makes approx equal sized chunks of a vector
}
\examples{
make_chunks(1:10, 3)
}
