% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fun.R
\name{make_fun}
\alias{make_fun}
\title{Makes a file "<name>.R", similar to usethis::use_r}
\usage{
make_fun(name, folder = here::here("R"))
}
\arguments{
\item{name}{name of function}

\item{folder}{folder}
}
\description{
Makes a file "<name>.R", similar to usethis::use_r
}
\examples{
make_fun('est_modus')
}
\keyword{function}
