% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grids.R
\name{make_grids}
\alias{make_grids}
\title{Makes 1km grids according to INSPIRE for a rectangle}
\usage{
make_grids(xmin, xmax, ymin, ymax)
}
\arguments{
\item{xmin}{Min x coordinate (in 1000)}

\item{xmax}{Max x coordinate (in 1000)}

\item{ymin}{Min y coordinate (in 1000)}

\item{ymax}{Max y coordinate (in 1000)}
}
\description{
Makes 1km grids according to INSPIRE for a rectangle
}
\examples{
make_grids(4010, 4015, 2050, 2060)
}
