% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hive_folder.R
\name{make_hive_folder}
\alias{make_hive_folder}
\title{Makes a hive_folder for arrow}
\usage{
make_hive_folder(basefolder, ..., pairs = list(...))
}
\arguments{
\item{basefolder}{Basic folder}

\item{...}{Key value pairs. Specify this OR pairs}

\item{pairs}{Named list of key values parirs; etc. day=01}
}
\description{
Makes a hive_folder for arrow
}
\examples{
make_folder("", year = 2015, month = "01", day = "15")
}
\keyword{hive_folder}
