% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_git.R
\name{prep_git}
\alias{prep_git}
\title{Prepares git in project folder}
\usage{
prep_git(
  target_folder = NULL,
  ignore = c(".Rproj.user/", ".Rhistory", ".DS_store"),
  github = TRUE,
  github_private = TRUE,
  repo_url = "https://github.com/matthiaskaeding/"
)
}
\arguments{
\item{target_folder}{target folder}

\item{ignore}{Character vector of files to ignore}

\item{github}{Create github repo?}

\item{github_private}{Should this be private (logical)?}
}
\description{
Mimics default github setup without readme
}
\keyword{git,}
\keyword{github}
\keyword{project,}
