% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_git_with_github.R
\name{prep_git_with_github}
\alias{prep_git_with_github}
\title{Prepares git in project folder}
\usage{
prep_git_with_github(
  target_folder,
  ignore = c("/.Rproj.user", "*.Rhistory", "DS_store")
)
}
\arguments{
\item{target_folder}{Path of target folder}

\item{ignore}{Character vector of files to ignore}
}
\description{
Mimics default github setup without readme
}
\examples{
prep_project("awesome_project")
}
\keyword{git,}
\keyword{github}
\keyword{project,}
