% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_project.R
\name{prep_project}
\alias{prep_project}
\title{Prepares project folder structure WITHIN a project}
\usage{
prep_project(
  target_folder,
  drake = TRUE,
  default_folders = c("R", "python", "data", "plots", "lit", "doc"),
  github
)
}
\arguments{
\item{drake}{use {drake}?}

\item{github}{Create github repo?}

\item{git}{git repo?}

\item{git_author}{Author in}

\item{ignore}{Input for gitignore}
}
\description{
Meant to be used after usethis::create_project
Sets up files, folders etc
}
\examples{
prepare_project()
}
\keyword{git,}
\keyword{github}
\keyword{project,}
