% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{rd}
\alias{rd}
\title{Run a drake project; unlock cache before
Simple helper to avoid the need to create run.R}
\usage{
rd()
}
\description{
Run a drake project; unlock cache before
Simple helper to avoid the need to create run.R
}
\keyword{drake}
