% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_windows_path.R
\name{rep_windows_path}
\alias{rep_windows_path}
\title{Converts a windows path from clipboard, turning backslash into forward slash}
\usage{
rep_windows_path()
}
\description{
Inspired by similar function from user Tom from stackoverflow
}
\keyword{windows}
