% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_windows_path_clip.R
\name{rep_windows_path_clip}
\alias{rep_windows_path_clip}
\title{Converts a windows path from clipboard, copy to clipboard, turning backslash into forward slash}
\usage{
rep_windows_path_clip()
}
\description{
Inspired by similar function from user Tom from stackoverflow
}
\keyword{windows}
