% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_load_fun_raw.R
\name{tar_load_fun_raw}
\alias{tar_load_fun_raw}
\title{targets::tar_loads all arguments of function
Common usecase: tar_load all objects while debugging a make_<object> function}
\usage{
tar_load_fun_raw(fun, envir = parent.frame())
}
\arguments{
\item{fun}{function (string), whose arguments we want to loadd}
}
\description{
targets::tar_loads all arguments of function
Common usecase: tar_load all objects while debugging a make_<object> function
}
\keyword{drake}
