% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cols.R
\name{add_cols}
\alias{add_cols}
\title{Adds cols from data.table SOURCE to data.table TARGET}
\usage{
add_cols(TARGET, SOURCE, on, cols)
}
\arguments{
\item{TARGET}{Target data.table}

\item{SOURCE}{Source data.table}
}
\description{
Only done when there if rows are uniquely identified by on
Returns TARGET invisibly
}
\examples{
TARGET = data.table(id = 1:10)
SOURCE = data.table(id = 1:5)[, x := rnorm(.N)][, y := rnorm(.N)]
add_cols(TARGET, SOURCE, on = "id", cols = c("x", "y"))
TARGET[]
}
\keyword{Update}
\keyword{data.table}
