% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_proj.R
\name{prep_proj}
\alias{prep_proj}
\title{Creates project folder.}
\usage{
prep_proj(
  target_folder,
  targets = TRUE,
  default_folders = c("R", "python", "data", "plots", "lit", "docs")
)
}
\arguments{
\item{target_folder}{Path of target folder}

\item{targets}{use targets packages?}

\item{default_folders}{List of folders to create}
}
\description{
Sets up files, folders etc. Inspired by dflow::use_dflow
}
\examples{
prep_project("awesome_project")
}
\keyword{drake}
\keyword{project,}
