% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_proj.R
\name{prep_proj}
\alias{prep_proj}
\title{Creates project folder.}
\usage{
prep_proj(
  target_folder,
  drake = TRUE,
  default_folders = c("R", "python", "data", "plots", "lit", "doc"),
  github
)
}
\arguments{
\item{target_folder}{Path of target folder}

\item{drake}{use drake?}

\item{default_folders}{List of folders to create}

\item{github}{Set up github page? If specified, must be must be either "private" or "public",}
}
\description{
Sets up files, folders etc. Inspired by dflow::use_dflow
}
\examples{
prep_project("awesome_project")
}
\keyword{drake}
\keyword{project,}
